+++ utils/stringManipulation/addHashToArrayString.jsx
@@ -1,3 +1,0 @@
-export default function stylingTagStringWithHash(content) {
-  return content.tags.map((tag) => `#${tag.tag}`).join(" ");
-}